Move files around for cleaner layout
Until we theme the CMS and the common templates as well as the LMS,
we won't create specific directories for them. This keeps the theme
includes in the main edX templates cleaner (no `lms/` prefix).

This commit also adds some cleanup changes to the header, like
using the `static.url()` helper and adjusting block names.

NOTE: this is incompatible with the existing changes on
test.class.stanford.edu and class.stanford.edu. DO NOT DEPLOY this
until the theming PR lands on edx-platform/master.