Default theme, no longer Stanford's theme
Previously this repo had the assets and styling info that Stanford used
for our own instance of Open edX, but we found that people were standing
up instances of Open edX that looked like our site!  No good.  So we've
now made the Stanford theme private and are providing this generic theme
to the community.

This is exactly the contents of the Stanford-Online/edx-default-theme
repo, but now here where it will be easier to find.

The Stanford theme is now in Stanford-Online/stanford-theme (private).